(PC-7565) : Move Favorite to pcapi.core.users
